refactor(debug_server): extract /navigate + /screen + nav target machinery to debug_server_nav (refs #332)
Wave 23b · 14th per-family extract.  All screen-navigation surfaces
move from debug_server.c into a new sibling module debug_server_nav.{c,h}:

  POST /navigate?screen=...  — LVGL-thread screen swap with 500 ms debounce
  GET  /screen               — current screen + overlay visibility (chat /
                                voice / settings) for the e2e harness

Shared infra moved with the family:
  - s_nav_target (last-requested screen) + s_navigating (re-entry guard)
  - tab5_debug_set_nav_target (PUBLIC setter — ui_chrome / ui_chat /
    ui_camera / ui_files / ui_notes / ui_settings / ui_wifi all call this
    on swipe-back / back-button / home-button paths to keep the harness's
    /screen view honest about the actually-loaded screen)
  - async_navigate (LVGL-thread callback that dismisses overlays + creates
    the target screen — handles 13 known target names, NULL-guards
    everything internally)

E2E verification on physical Tab5 192.168.1.90:
  - story_smoke 14/14 in 90s — every step calls /navigate via the harness

debug_server.c LOC: 2,098 → ~1,830 (-268).
debug_server_nav.c: 0 → 234.
